StageManagerに機能を移行
