+++ Assets/Childs Garden/06_Script/03_GameScene/01_MainManager/StageManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c378749de8884024e9e8dd557e38154a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
